pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lazy-take-notes"
version = "0.1.0"
description = "TUI for real-time transcription and AI-assisted note-taking"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "textual>=0.89",
    "pywhispercpp>=1.4.0",
    "ollama>=0.6",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "sounddevice>=0.5.0",
    "numpy>=1.24.0",
    "huggingface_hub>=0.20.0",
    "pyperclip>=1.8",
]

[project.scripts]
ltn = "lazy_take_notes.l4_frameworks_and_drivers.cli:cli"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "textual-dev>=1.0",
    "import-linter>=2.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/lazy_take_notes"]
artifacts = [
    "src/lazy_take_notes/_native/bin/coreaudio-tap",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "W",    # pycodestyle warnings
    "I",    # isort
    "B",    # flake8-bugbear
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "S",    # flake8-bandit
    "PLC",  # pylint convention
]
preview = true

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S"]
"test_*" = ["S"]

[tool.ruff.format]
quote-style = "single"

[tool.ty.environment]
python-version = "3.11"
python = ".venv"

[tool.importlinter]
root_packages = ["lazy_take_notes"]
include_external_packages = true

[[tool.importlinter.contracts]]
name = "L1 entities must not import L2, L3, L4, or framework libs"
type = "forbidden"
source_modules = ["lazy_take_notes.l1_entities"]
forbidden_modules = [
    "lazy_take_notes.l2_use_cases",
    "lazy_take_notes.l3_interface_adapters",
    "lazy_take_notes.l4_frameworks_and_drivers",
    "textual",
    "ollama",
    "pywhispercpp",
    "yaml",
    "click",
    "sounddevice",
]

[[tool.importlinter.contracts]]
name = "L2 use cases must not import L3, L4, or framework libs"
type = "forbidden"
source_modules = ["lazy_take_notes.l2_use_cases"]
forbidden_modules = [
    "lazy_take_notes.l3_interface_adapters",
    "lazy_take_notes.l4_frameworks_and_drivers",
    "textual",
    "ollama",
    "pywhispercpp",
    "yaml",
    "click",
    "sounddevice",
]

[[tool.importlinter.contracts]]
name = "L3 adapters must not import L4"
type = "forbidden"
source_modules = ["lazy_take_notes.l3_interface_adapters"]
forbidden_modules = [
    "lazy_take_notes.l4_frameworks_and_drivers",
]

[[tool.importlinter.contracts]]
name = "L3 adapters must not import textual"
type = "forbidden"
source_modules = ["lazy_take_notes.l3_interface_adapters"]
forbidden_modules = [
    "textual",
]
